Fixed tests and added new behaviour. Ref.: #221
I) Tests work now for a generic event
II) You could create Events with a playercount per team
III) For single player events a team with playercounts of 1 will be created automatically.
IV) It is not possible not create events where the min teamsize is bigger than the maximum teamsize.